+++ CHANGELOG.md
@@ -9,6 +9,72 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Added
+
+- Add `addFixed(child:)` and `addFlexible(child:)` methods to `StackElement` for adding children with a grow & shrink priority of 0.0 and 1.0 respectively.
+
+### Removed
+
+### Changed
+
+### Deprecated
+
+### Security
+
+### Documentation
+
+### Misc
+
+# Past Releases
+
+## [0.14.0] - 2020-08-12
+
+### Added
+
+- Add `textColor` property on TextField ([#133](https://github.com/square/Blueprint/pull/133)).
+- Add the `windowSize` environment key. ([#134])
+
+- Add `GeometryReader`. ([#135])
+
+  This element allow you to compose elements whose contents depend on the amount of space available.
+
+  Here is an example that dynamically chooses an image based on the width available:
+
+  ```swift
+  GeometryReader { (geometry) -> Element in
+      let image: UIImage
+      switch geometry.constraint.width.maximum {
+      case ..<100:
+          image = UIImage(named: "small")!
+      case 100..<500:
+          image = UIImage(named: "medium")!
+      default:
+          image = UIImage(named: "large")!
+      }
+      return Image(image: image)
+  }
+  ```
+
+### Changed
+
+- Default `ScrollView.delaysContentTouches` to `true` ([#132](https://github.com/square/Blueprint/pull/132))
+
+### Misc
+
+- Set an explicit shadow path on `Box` ([#137](https://github.com/square/Blueprint/pull/137))
+
+## [0.13.1] - 2020-07-30
+
+### Added
+
+- Introduce `AccessibilityContainer` element for wrapping an element with multiple sub-elements that should be in a voice over container.
+
+- Add `font` property on TextField ([#127](https://github.com/square/Blueprint/pull/127)).
+
+## [0.13.0] - 2020-07-20
+
+### Fixed
+
 - [Update the scroll indicator inset](https://github.com/square/Blueprint/pull/117) when adjusting the content inset.
 
 - `Label` & `AttributedLabel` use an internal `UILabel` for measurement. This fixes measurement when there is a line limit set. However, it also means that the screen scale cannot be specified and is always assumed to be `UIScreen.main.scale`. These elements may not be measured correctly if they are placed on a screen other than `UIScreen.main`. ([#120])
@@ -43,23 +109,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `isAccessibilityElement` to `Label` and `AttributedLabel`. ([#120])
 - Add `lineHeight` to `Label` for specifying custom line heights. `AttributedLabel` has a `textRectOffset` property to support this. ([#120])
 
-### Removed
-
 ### Changed
 
 - [Update Demo app](https://github.com/square/Blueprint/pull/116) to support more demo screen types.
 
-### Deprecated
-
-### Security
-
-### Documentation
-
-### Misc
-
-# Past Releases
-
-## [0.12.2] - 06-08-2020
+## [0.12.2] - 2020-06-08
 
 ### Fixed
 
@@ -69,13 +123,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add [delaysContentTouches](https://github.com/square/Blueprint/pull/109) to the `ScrollView` element.
 
-## [0.12.1] - 06-05-2020
+## [0.12.1] - 2020-06-05
 
 ### Fixed
 
 - Use default environment when [measuring `BlueprintView`](https://github.com/square/Blueprint/pull/107).
 
-## [0.12.0] - 06-04-2020
+## [0.12.0] - 2020-06-04
 
 ### Fixed
 
@@ -306,7 +360,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First stable release.
 
-[main]: https://github.com/square/Blueprint/compare/0.12.2...HEAD
+[main]: https://github.com/square/Blueprint/compare/0.14.0...HEAD
+[0.14.0]: https://github.com/square/Blueprint/compare/0.14.0...0.13.1
+[0.13.1]: https://github.com/square/Blueprint/compare/0.13.1...0.13.0
+[0.13.0]: https://github.com/square/Blueprint/compare/0.13.0...0.12.2
 [0.12.2]: https://github.com/square/Blueprint/compare/0.12.1...0.12.2
 [0.12.1]: https://github.com/square/Blueprint/compare/0.12.0...0.12.1
 [0.12.0]: https://github.com/square/Blueprint/compare/0.11.0...0.12.0
@@ -323,6 +380,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.1]: https://github.com/square/Blueprint/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/square/Blueprint/compare/0.2.2...0.3.0
 [0.2.2]: https://github.com/square/Blueprint/releases/tag/0.2.2
+[#135]: https://github.com/square/Blueprint/pull/135
+[#134]: https://github.com/square/Blueprint/pull/134
 [#120]: https://github.com/square/Blueprint/pull/120
 [#102]: https://github.com/square/Blueprint/pull/102
 [#101]: https://github.com/square/Blueprint/pull/101